+++ generator/proto/_utils.py
